@@ -6,7 +6,7 @@ authors = ["timoniq <tesseradecades@mail.ru>"]
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.11"
 aiohttp = "^3.8.1"
 vbml = "^1.1.post1"
 choicelib = "^0.1.5"
