@@ -1,6 +1,7 @@
-import typing
-
-
 class APIError(BaseException):
-    def __init__(self, code: int, error: typing.Optional[str] = None):
+    def __init__(self, code: int, error: str | None = None):
         self.code, self.error = code, error
+
+
+class InvalidTokenError(BaseException):
+    ...
