@@ -1,7 +1,7 @@
+from telegrinder.bot.dispatch.middleware.abc import ABCMiddleware
 from .abc import ABCView
 from telegrinder.bot.dispatch.handler import ABCHandler, FuncHandler
 from telegrinder.bot.dispatch.waiter import Waiter
-from telegrinder.bot.dispatch.middleware import ABCMiddleware
 from telegrinder.bot.rules import ABCRule
 from telegrinder.bot.cute_types import InlineQueryCute
 from telegrinder.api.abc import ABCAPI
@@ -14,23 +14,20 @@ import typing
 class InlineQueryView(ABCView, WithWaiter[str, InlineQueryCute]):
     def __init__(self):
         self.auto_rules: list[ABCRule] = []
-        self.handlers: typing.List[ABCHandler[InlineQueryCute]] = []
-        self.middlewares: typing.List[ABCMiddleware[InlineQueryCute]] = []
-        self.short_waiters: typing.Dict[str, Waiter] = {}
+        self.handlers: list[ABCHandler[InlineQueryCute]] = []
+        self.middlewares: list[ABCMiddleware[InlineQueryCute]] = []
+        self.short_waiters: dict[str, Waiter] = {}
 
     def __call__(self, *rules: ABCRule, is_blocking: bool = True):
         def wrapper(func: typing.Callable[..., typing.Coroutine]):
             self.handlers.append(
-                FuncHandler(func, [*self.auto_rules, *rules], is_blocking, dataclass=None)
+                FuncHandler(
+                    func, [*self.auto_rules, *rules], is_blocking, dataclass=None
+                )
             )
             return func
 
         return wrapper
-
-    def load(self, external: "InlineQueryView"):
-        self.handlers.extend(external.handlers)
-        self.middlewares.extend(external.middlewares)
-        external.short_waiters = self.short_waiters
 
     async def check(self, event: Update) -> bool:
         return bool(event.inline_query)
@@ -39,7 +36,7 @@ class InlineQueryView(ABCView, WithWaiter[str, InlineQueryCute]):
         query = InlineQueryCute(**event.inline_query.to_dict(), api=api)
 
         if await process_waiters(
-            self.short_waiters, query.id, query, event, query.answer
+            api, self.short_waiters, query.id, query, event, query.answer
         ):
             return
 
