@@ -1,28 +1,23 @@
 from .abc import Message, patcher
-from .text import ABCTextMessageRule
-import typing
+from .text import TextMessageRule
 import vbml
 
-PatternLike = typing.Union[str, vbml.Pattern]
+PatternLike = str | vbml.Pattern
 
 
-def check_string(
-    patterns: typing.Union[PatternLike, typing.List[PatternLike]], s: str, ctx: dict
-) -> bool:
-    if s is None:
-        return False
+def check_string(patterns: list[PatternLike], s: str, ctx: dict) -> bool:
     for pattern in patterns:
-        response = patcher.check(pattern, s)
-        if response in (False, None):
-            continue
-        elif isinstance(response, dict):
-            ctx.update(response)
+        match patcher.check(pattern, s):
+            case None | False:
+                continue
+            case {**response}:
+                ctx |= response
         return True
     return False
 
 
-class Markup(ABCTextMessageRule):
-    def __init__(self, patterns: typing.Union[PatternLike, typing.List[PatternLike]]):
+class Markup(TextMessageRule):
+    def __init__(self, patterns: PatternLike | list[PatternLike]):
         if not isinstance(patterns, list):
             patterns = [patterns]
         self.patterns = [
