@@ -1,26 +1,23 @@
-from .abc import ABC, ABCMessageRule, Message
-import typing
+from .abc import ABC, MessageRule, Message
 
 
-class HasText(ABCMessageRule):
+class HasText(MessageRule):
     async def check(self, message: Message, ctx: dict) -> bool:
         return bool(message.text)
 
 
-class ABCTextMessageRule(ABCMessageRule, ABC, require=[HasText()]):
+class TextMessageRule(MessageRule, ABC, require=[HasText()]):
     pass
 
 
-class Text(ABCTextMessageRule):
-    def __init__(
-        self, texts: typing.Union[str, typing.List[str]], ignore_case: bool = False
-    ):
+class Text(TextMessageRule):
+    def __init__(self, texts: str | list[str], ignore_case: bool = False):
         if not isinstance(texts, list):
             texts = [texts]
-        self.texts = texts
+        self.texts = texts if not ignore_case else list(map(str.lower, self.texts))
         self.ignore_case = ignore_case
 
     async def check(self, message: Message, ctx: dict) -> bool:
-        if self.ignore_case:
-            return message.text.lower() in list(map(str.lower, self.texts))
-        return message.text in self.texts
+        return (
+            message.text if not self.ignore_case else message.text.lower()
+        ) in self.texts
