@@ -1,35 +1,29 @@
 from telegrinder import Telegrinder, API, Token, Message
 from telegrinder.rules import Text
-from telegrinder.tools import MarkdownFormatter, HTMLFormatter
+from telegrinder.tools import HTMLFormatter, bold, italic, mention
 import logging
 
 api = API(token=Token.from_env())
 bot = Telegrinder(api)
 logging.basicConfig(level=logging.INFO)
-formatter = MarkdownFormatter
 
 
 @bot.on.message(Text("/formatting"))
 async def formatting(m: Message):
     await m.answer(
-        formatter("bold italic").bold().italic()
-        + " and "
-        + formatter("strike link")
-        .link("https://github.com/timoniq/telegrinder")
-        .strike(),
-        parse_mode=formatter.PARSE_MODE,
+        bold(italic("bold italic text!")), parse_mode=HTMLFormatter.PARSE_MODE
     )
     await m.answer(
-        "this is " + formatter("mention").mention(m.from_.id),
-        parse_mode=formatter.PARSE_MODE,
+        "python library 'telegrinder' - "
+        + HTMLFormatter(
+            "{:bold} for effective and reliable telegram {:bold+italic} {:underline}!"
+        ).format("Framework", "bot", "building"),
+        parse_mode=HTMLFormatter.PARSE_MODE,
     )
-
-
-@bot.on.message(Text("/change"))
-async def change(m: Message):
-    global formatter
-    formatter = HTMLFormatter if formatter == MarkdownFormatter else MarkdownFormatter
-    await m.answer(f"Formatter changed to {formatter.__name__}")
+    await m.answer(
+        "this is " + mention(m.from_.first_name, m.from_.id),
+        parse_mode=HTMLFormatter.PARSE_MODE,
+    )
 
 
 bot.run_forever()
