@@ -1,4 +1,4 @@
 from .abc import ABCAPI, APIError, Token
 from .api import API
 from .response import APIResponse
-from .error import APIError
+from .error import APIError, InvalidTokenError
